+++ CHANGELOG.md
@@ -8,5 +8,6 @@ The change log begins with version 2.1.30.
 
 ### Changed
 
+- Add screen protection feature
 - Removed support for creating shortcuts on the desktop
 - The maximum number of tags that can be created has been updated to 25.